Move formatArchiveSeoData to seo folder
