+++ functions/wordpress/seo/formatArchiveSeoData.js
